refactor: create dedicated ConversationMenu
