+++ front/components/assistant/conversation/ConversationTitle.tsx
@@ -1,32 +1,19 @@
 import {
   Button,
   CheckIcon,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuTrigger,
   Input,
-  LinkIcon,
-  MoreIcon,
   PencilSquareIcon,
-  TrashIcon,
   XMarkIcon,
 } from "@dust-tt/sparkle";
-import { useRouter } from "next/router";
 import type { MouseEvent } from "react";
 import { useCallback, useRef, useState } from "react";
 import { useSWRConfig } from "swr";
 
 import { ConversationFilesPopover } from "@app/components/assistant/conversation/ConversationFilesPopover";
+import { ConversationMenu } from "@app/components/assistant/conversation/ConversationMenu";
 import { useConversationsNavigation } from "@app/components/assistant/conversation/ConversationsNavigationProvider";
-import { DeleteConversationsDialog } from "@app/components/assistant/conversation/DeleteConversationsDialog";
 import { AppLayoutTitle } from "@app/components/sparkle/AppLayoutTitle";
-import { useSendNotification } from "@app/hooks/useNotification";
-import {
-  useConversation,
-  useDeleteConversation,
-} from "@app/lib/swr/conversations";
+import { useConversation } from "@app/lib/swr/conversations";
 import type { WorkspaceType } from "@app/types";
 
 export function ConversationTitle({
@@ -37,38 +24,18 @@ export function ConversationTitle({
   baseUrl: string;
 }) {
   const { mutate } = useSWRConfig();
-  const router = useRouter();
   const { activeConversationId } = useConversationsNavigation();
-
-  const sendNotification = useSendNotification();
 
   const { conversation } = useConversation({
     conversationId: activeConversationId,
     workspaceId: owner.sId,
   });
 
-  const shareLink = `${baseUrl}/w/${owner.sId}/assistant/${activeConversationId}`;
-
-  const doDelete = useDeleteConversation(owner);
-
-  const onDelete = useCallback(async () => {
-    const res = await doDelete(conversation);
-    if (res) {
-      void router.push(`/w/${owner.sId}/assistant/new`);
-    }
-  }, [conversation, doDelete, owner.sId, router]);
-
   const [isEditingTitle, setIsEditingTitle] = useState<boolean>(false);
-  const [showDeleteDialog, setShowDeleteDialog] = useState<boolean>(false);
   const [editedTitle, setEditedTitle] = useState<string>("");
 
   const titleInputFocused = useRef(false);
   const saveButtonFocused = useRef(false);
-
-  const copyConversationLink = useCallback(async () => {
-    await navigator.clipboard.writeText(shareLink || "");
-    sendNotification({ type: "success", title: "Link copied !" });
-  }, [shareLink, sendNotification]);
 
   const onTitleChange = useCallback(
     async (title: string) => {
@@ -104,119 +71,93 @@ export function ConversationTitle({
   }
 
   return (
-    <>
-      <DeleteConversationsDialog
-        isOpen={showDeleteDialog}
-        type="selection"
-        selectedCount={1}
-        onClose={() => setShowDeleteDialog(false)}
-        onDelete={() => {
-          setShowDeleteDialog(false);
-          void onDelete();
-        }}
-      />
-      <AppLayoutTitle>
-        <div className="grid h-full min-w-0 max-w-full grid-cols-[1fr,auto] items-center gap-4">
-          <div className="flex min-w-0 flex-row items-center gap-4 text-primary dark:text-primary-night">
-            {!isEditingTitle ? (
-              <div className="dd-privacy-mask min-w-0 overflow-hidden truncate text-sm font-normal">
-                {conversation?.title || ""}
-              </div>
-            ) : (
-              <div className="w-[84%]">
-                <Input
-                  value={editedTitle}
-                  onChange={(e) => setEditedTitle(e.target.value)}
-                  onFocus={() => (titleInputFocused.current = true)}
-                  onBlur={() => {
-                    setTimeout(() => {
-                      if (!saveButtonFocused.current) {
-                        setIsEditingTitle(false);
-                      }
-                      titleInputFocused.current = false;
-                    }, 0);
-                  }}
-                  onKeyDown={(e) => {
-                    if (e.key === "Enter") {
-                      return onTitleChange(editedTitle);
+    <AppLayoutTitle>
+      <div className="grid h-full min-w-0 max-w-full grid-cols-[1fr,auto] items-center gap-4">
+        <div className="flex min-w-0 flex-row items-center gap-4 text-primary dark:text-primary-night">
+          {!isEditingTitle ? (
+            <div className="dd-privacy-mask min-w-0 overflow-hidden truncate text-sm font-normal">
+              {conversation?.title || ""}
+            </div>
+          ) : (
+            <div className="w-[84%]">
+              <Input
+                value={editedTitle}
+                onChange={(e) => setEditedTitle(e.target.value)}
+                onFocus={() => (titleInputFocused.current = true)}
+                onBlur={() => {
+                  setTimeout(() => {
+                    if (!saveButtonFocused.current) {
+                      setIsEditingTitle(false);
                     }
-                  }}
-                  autoFocus
-                />
-              </div>
-            )}
-
-            {isEditingTitle ? (
-              <div className="flex flex-row gap-2">
-                <div
-                  onClick={(e: MouseEvent<HTMLDivElement>) => {
-                    e.preventDefault();
-                    return onTitleChange(editedTitle);
-                  }}
-                  // See comment on the input above.
-                  onFocus={() => (saveButtonFocused.current = true)}
-                  onBlur={() => {
-                    setTimeout(() => {
-                      if (!titleInputFocused.current) {
-                        setIsEditingTitle(false);
-                      }
-                      saveButtonFocused.current = false;
-                    }, 0);
-                  }}
-                  className="flex items-center"
-                >
-                  <Button size="mini" icon={CheckIcon} variant="primary" />
-                </div>
-                <Button
-                  icon={XMarkIcon}
-                  size="mini"
-                  onClick={() => {
-                    setIsEditingTitle(false);
-                    setEditedTitle("");
-                  }}
-                  variant="outline"
-                />
-              </div>
-            ) : (
-              <Button
-                icon={PencilSquareIcon}
-                onClick={() => {
-                  setEditedTitle(conversation?.title || "");
-                  setIsEditingTitle(true);
+                    titleInputFocused.current = false;
+                  }, 0);
                 }}
-                size="mini"
-                variant="ghost-primary"
+                onKeyDown={(e) => {
+                  if (e.key === "Enter") {
+                    return onTitleChange(editedTitle);
+                  }
+                }}
+                autoFocus
               />
-            )}
-          </div>
-          <div className="flex items-center gap-2">
-            <ConversationFilesPopover
-              conversationId={activeConversationId}
-              owner={owner}
+            </div>
+          )}
+
+          {isEditingTitle ? (
+            <div className="flex flex-row gap-2">
+              <div
+                onClick={(e: MouseEvent<HTMLDivElement>) => {
+                  e.preventDefault();
+                  return onTitleChange(editedTitle);
+                }}
+                // See comment on the input above.
+                onFocus={() => (saveButtonFocused.current = true)}
+                onBlur={() => {
+                  setTimeout(() => {
+                    if (!titleInputFocused.current) {
+                      setIsEditingTitle(false);
+                    }
+                    saveButtonFocused.current = false;
+                  }, 0);
+                }}
+                className="flex items-center"
+              >
+                <Button size="mini" icon={CheckIcon} variant="primary" />
+              </div>
+              <Button
+                icon={XMarkIcon}
+                size="mini"
+                onClick={() => {
+                  setIsEditingTitle(false);
+                  setEditedTitle("");
+                }}
+                variant="outline"
+              />
+            </div>
+          ) : (
+            <Button
+              icon={PencilSquareIcon}
+              onClick={() => {
+                setEditedTitle(conversation?.title || "");
+                setIsEditingTitle(true);
+              }}
+              size="mini"
+              variant="ghost-primary"
             />
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button size="sm" variant="ghost" icon={MoreIcon} />
-              </DropdownMenuTrigger>
-              <DropdownMenuContent>
-                <DropdownMenuLabel>Conversation</DropdownMenuLabel>
-                <DropdownMenuItem
-                  label="Delete"
-                  onClick={() => setShowDeleteDialog(true)}
-                  icon={TrashIcon}
-                  variant="warning"
-                />
-                <DropdownMenuLabel>Share the conversation</DropdownMenuLabel>
-                <DropdownMenuItem
-                  label="Copy the link"
-                  onClick={copyConversationLink}
-                  icon={LinkIcon}
-                />
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
+          )}
         </div>
-      </AppLayoutTitle>
-    </>
+        <div className="flex items-center gap-2">
+          <ConversationFilesPopover
+            conversationId={activeConversationId}
+            owner={owner}
+          />
+          <ConversationMenu
+            activeConversationId={activeConversationId}
+            conversation={conversation}
+            baseUrl={baseUrl}
+            owner={owner}
+          />
+        </div>
+      </div>
+    </AppLayoutTitle>
   );
 }
